requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
kfp
pandas
scikit-learn
joblib
